fix(experimental): resolve lint errors introduced by this branch
Fix all ESLint errors from the experimental-api-flag branch: break
import/no-cycle by moving MULTI_SOURCE_EVENT_NAMES to a new data file,
extract RpcEventRegistry to its own file (max-classes-per-file), fix
default-param-last in registerEngine/registerEngineByType, add
expectTypeOf to vitest assertFunctionNames, and fix no-use-before-define,
no-null, no-unnecessary-type-assertion, and various test-file issues.
