+++ src/shared/services/__tests__/network.service.shared-events.test.ts
@@ -5,6 +5,7 @@ import {
   getNetworkEvent,
 } from '@shared/services/network.service';
 import type { MultiSourceNetworkEvents } from 'papi-shared-types';
+import type { PlatformEvent } from 'platform-bible-utils';
 
 describe('MULTI_SOURCE_EVENT_NAMES stays in sync with MultiSourceNetworkEvents', () => {
   it('contains every key of MultiSourceNetworkEvents', () => {
@@ -14,18 +15,18 @@ describe('MULTI_SOURCE_EVENT_NAMES stays in sync with MultiSourceNetworkEvents',
       'object:onDidDisposeNetworkObject',
       'shared-store:change',
     ];
-    for (const name of required) expect(MULTI_SOURCE_EVENT_NAMES.has(name)).toBe(true);
+    required.forEach((name) => expect(MULTI_SOURCE_EVENT_NAMES.has(name)).toBe(true));
   });
 
   it('contains no names absent from MultiSourceNetworkEvents', () => {
     type AllowedKeys = keyof MultiSourceNetworkEvents;
-    for (const name of MULTI_SOURCE_EVENT_NAMES) {
+    Array.from(MULTI_SOURCE_EVENT_NAMES).forEach((name) => {
       // The cast verifies the runtime constant only contains keys MultiSourceNetworkEvents declares.
       // If a new entry is added to MULTI_SOURCE_EVENT_NAMES without adding it to
       // MultiSourceNetworkEvents, this assignment is a type error and the test fails to compile.
       const typed: AllowedKeys = name;
       expect(typeof typed).toBe('string');
-    }
+    });
   });
 });
 
@@ -44,13 +45,22 @@ describe.skip('createNetworkEventEmitterAsync', () => {
       // The 'as never' cast bypasses the NetworkEventTypes constraint at the call site for a name
       // that won't have been declared in this test's compilation unit. The runtime accepts any
       // string; the typed constraint is enforced at production call sites by NetworkEventTypes.
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
       const emitter = await createNetworkEventEmitterAsync('myExt.testEvent' as never);
       let received: unknown;
+      // 'as never' bypasses the NetworkEventTypes constraint — same reason as above.
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
       getNetworkEvent('myExt.testEvent' as never)((event) => {
         received = event;
       });
+      // 'as never' bypasses the typed payload constraint for an ad-hoc test event.
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
       emitter.emit({ value: 42 } as never);
-      await new Promise((r) => setTimeout(r, 10));
+      await new Promise<void>((resolve) => {
+        setTimeout(resolve, 10);
+      });
+      // Cast is needed: `received` is typed as `unknown`; narrowing to the expected shape.
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
       expect((received as { value: number }).value).toBe(42);
     } catch (err) {
       if (
@@ -66,7 +76,11 @@ describe.skip('createNetworkEventEmitterAsync', () => {
 
   it('rejects a second registration of the same single-source event name within one process', async () => {
     try {
+      // 'as never' bypasses the NetworkEventTypes constraint for a test-only event name.
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
       await createNetworkEventEmitterAsync('myExt.unique' as never);
+      // 'as never' bypasses the NetworkEventTypes constraint for a test-only event name.
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
       await expect(createNetworkEventEmitterAsync('myExt.unique' as never)).rejects.toThrow();
     } catch (err) {
       if (
@@ -80,21 +94,18 @@ describe.skip('createNetworkEventEmitterAsync', () => {
   });
 });
 
-import { getNetworkEvent as gne } from '@shared/services/network.service';
-import type { PlatformEvent } from 'platform-bible-utils';
-
 describe('getNetworkEvent overloads', () => {
   it('typed call infers payload from NetworkEvents', () => {
     // Compile-time test — if this compiles, the typed overload exists.
-    const ev = gne('object:onDidCreateNetworkObject');
+    const ev = getNetworkEvent('object:onDidCreateNetworkObject');
     // ev should be PlatformEvent<NetworkObjectDetails> — verify by satisfying the structural type.
     const checked: PlatformEvent<unknown> = ev;
-    void checked;
+    expect(checked).toBeDefined();
   });
 
   it('explicit <T> call still resolves (deprecated overload)', () => {
-    const ev = gne<{ foo: string }>('arbitraryName');
+    const ev = getNetworkEvent<{ foo: string }>('arbitraryName');
     const checked: PlatformEvent<{ foo: string }> = ev;
-    void checked;
+    expect(checked).toBeDefined();
   });
 });
