+++ src/shared/services/__tests__/project-data-provider.service.test.ts
@@ -1,4 +1,4 @@
-import { describe, it } from 'vitest';
+import { describe, it, expect } from 'vitest';
 import { registerProjectDataProviderEngineFactory } from '@shared/services/project-data-provider.service';
 import type { IProjectDataProviderEngineFactory } from '@shared/models/project-data-provider-engine-factory.model';
 
@@ -7,26 +7,29 @@ describe('registerProjectDataProviderEngineFactory — attributes + documentatio
     // Compile-time only: verify the function signature.
     type Sig = Parameters<typeof registerProjectDataProviderEngineFactory>;
     // Should have 5 parameters: id, projectInterfaces, factory, attributes?, documentation?
-    void (null as unknown as Sig);
+    const sigCheck: Sig | undefined = undefined;
+    expect(sigCheck).toBeUndefined();
   });
 
   it('factory can return either engine or envelope (compile-time)', () => {
     // Verifies the return shape augmentation compiles.
     const factoryEngine: IProjectDataProviderEngineFactory<['platform.base']> = {
       getAvailableProjects: async () => [],
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      createProjectDataProviderEngine: async () => ({}) as any, // raw engine
+      // The cast is needed to satisfy the generic engine type in a test context without a real engine
+      // eslint-disable-next-line no-type-assertion/no-type-assertion
+      createProjectDataProviderEngine: async () => ({}) as never, // raw engine
     };
     const factoryEnvelope: IProjectDataProviderEngineFactory<['platform.base']> = {
       getAvailableProjects: async () => [],
       createProjectDataProviderEngine: async () => ({
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        projectDataProviderEngine: {} as any,
+        // The cast is needed to satisfy the generic engine type in a test context without a real engine
+        // eslint-disable-next-line no-type-assertion/no-type-assertion
+        projectDataProviderEngine: {} as never,
         attributes: { extra: 1 },
         documentation: { 'x-experimental': true },
       }),
     };
-    void factoryEngine;
-    void factoryEnvelope;
+    expect(factoryEngine).toBeDefined();
+    expect(factoryEnvelope).toBeDefined();
   });
 });
