+++ src/shared/services/network-object.service.ts
@@ -45,16 +45,24 @@ const initialize = (): Promise<void> => {
     // TODO: Might be best to make a singleton or something
     await networkService.initialize();
 
+    // `initialize` is only called after module evaluation is complete, so all module-level
+    // variables below are already defined by the time this async body runs.
+    // eslint-disable-next-line @typescript-eslint/no-use-before-define
     onDidCreateNetworkObjectEmitter = await networkService.createNetworkEventEmitterAsync(
       'object:onDidCreateNetworkObject',
     );
 
+    // `initialize` runs after module evaluation; the emitter variable is defined later in the module.
+    // eslint-disable-next-line @typescript-eslint/no-use-before-define
     onDidDisposeNetworkObjectEmitter = await networkService.createNetworkEventEmitterAsync(
       'object:onDidDisposeNetworkObject',
     );
 
     // Subscribe to the dispose event to clean up local and remote network object registrations
+    // eslint-disable-next-line @typescript-eslint/no-use-before-define
     onDidDisposeNetworkObject((id: string) => {
+      // networkObjectRegistrations is defined later in the module; safe at runtime since initialize runs after module eval.
+      // eslint-disable-next-line @typescript-eslint/no-use-before-define
       const networkObjectRegistration = networkObjectRegistrations.get(id);
 
       if (networkObjectRegistration) {
@@ -68,6 +76,7 @@ const initialize = (): Promise<void> => {
         networkObjectRegistration.revokeProxy();
 
         // Dispose of the network object registration
+        // eslint-disable-next-line @typescript-eslint/no-use-before-define
         networkObjectRegistrations.delete(id);
       }
     });
